Fix: Display freezing after 30 seconds
Switched the "end" time value from int to unsigned long (the value type of millis()'s result), as the amount of milliseconds passed after starting the Arduino fills an int in ~30 seconds

+++ Matrix/Matrix.ino
@@ -66,8 +66,8 @@ void generateMatrix(int input, bool matrix[8][8]) {
 }
 
 
-void display(bool matrix[8][8], int delayTime) {
-  int end = millis() + delayTime;
+void display(bool matrix[8][8], unsigned long delayTime) {
+  unsigned long end = millis() + delayTime;
   while (end >= millis()) {
     for (int rowPin = 0; rowPin < 8; rowPin++) {
       turnOff();
